added support for locally signed transactions in submitter
