+++ crates/e2e/tests/e2e/onchain_settlement.rs
@@ -20,7 +20,7 @@ use solver::{
     liquidity::uniswap_v2::UniswapLikeLiquidity,
     liquidity_collector::LiquidityCollector,
     metrics::NoopMetrics,
-    settlement_submission::{custom_nodes_api::CustomNodesApi, SolutionSubmitter, StrategyArgs},
+    settlement_submission::{submitter::custom_nodes_api::CustomNodesApi, SolutionSubmitter, StrategyArgs},
 };
 use std::{sync::Arc, time::Duration};
 use web3::signing::SecretKeyRef;
